requirements.txt
airsim>=1.8.1
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
opencv-python>=4.5.0
Pillow>=8.3.0
PyYAML>=5.4.0
transforms3d>=0.3.1
open3d>=0.13.0
PyQt5>=5.15.0
keyboard>=0.13.5
pyautogui>=0.9.52
psutil>=5.8.0
msgpack-rpc-python>=0.4.1
tqdm>=4.61.0
pandas>=1.3.0
scikit-learn>=0.24.0